+++ src/timePicker/index.unit.tsx
@@ -89,4 +89,11 @@ describe('<TimePicker />', () => {
       })
     })
   })
+
+  describe('#timeStart', () => {
+    it('Can set a time start', () => {
+      const wrapper = shallow(<TimePicker {...defaultProps} minuteStep={240} timeStart="08:00" />)
+      expect(Object.keys(wrapper.state('steps'))).toEqual(['08:00', '12:00', '16:00', '20:00'])
+    })
+  })
 })